+++ src/orthofinder_results/__init__.py
@@ -1,4 +1,4 @@
 """Generic interrogation and publication of OrthoFinder results."""
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 __schema_version__ = 2